Draft specification, make breaking change to file header
The file header has been changed to include XXHash64 of the
file header itself and XXHash64 values of the type headers.
This reduces the number of available types from 128 to 42, which
is still a large number of types (types may be increased
at the cost of allocating more than one 512 byte sector).
However, the tradeoff is well worth it as the headers are validated
upon opening to ensure they are not corrupt (previously, corruption
was only detected by invalid/overlapping sector allocations).