Add analyse operation for fast checking regionfile allocations
Retrieves stats so that fragmentation can be determined
as well as comparison of 4096 allocated sectors vs 512 byte
sectors. Since it does not do compression/decompression tests,
it should be faster.

Additionally, add in support for reading Vanilla oversized data
in RegionFile. This adds support for oversized vanilla data for
all tests except analyse, which will not calculate oversized
sector counts - as it is only concerned with internal allocations.