Add simplified ZSTD streams for usage in SectorFile
The old zstd streams were mainly a playground of ideas and are not
really fully tested. There may still be merit in the idea of merging
the ZSTD stream and datainput/dataoutput to reduce indirection
in DataInputStream/DataOutputStream, but that is something for
another day.

Re-do the compression tests in the specification using the new
ZSTD streams, which show a minor improvement in compression time.
Note that the compression ratios remain the same, as it is performed
on the same data.